requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
python-dotenv
razorpay
requests
aiohttp
asyncio